+++ Plugins/ActorPools/Source/NexusActorPools/Public/NPooledActor.h
@@ -14,6 +14,7 @@
 UCLASS(ClassGroup = "NEXUS", DisplayName = "Pooled Actor", Meta = (ShortTooltip = "An actor base which support utilizing an NActorPool."))
 class NEXUSACTORPOOLS_API ANPooledActor : public AActor, public INActorPoolItem
 {
+	DECLARE_DYNAMIC_MULTICAST_DELEGATE(FOnActorPoolDelegate);
 	GENERATED_BODY()
 
 	// ReSharper disable once CppUE4CodingStandardNamingViolationWarning
@@ -21,4 +22,41 @@ class NEXUSACTORPOOLS_API ANPooledActor : public AActor, public INActorPoolItem
 	{
 		ReturnToActorPool();
 	}
+	
+	virtual void OnCreatedByActorPool() override
+	{
+		INActorPoolItem::OnCreatedByActorPool();
+		OnCreatedByActorPoolEvent.Broadcast();
+	};
+
+	virtual void OnDestroyedByActorPool() override
+	{
+		INActorPoolItem::OnDestroyedByActorPool();
+		OnDestroyedByActorPoolEvent.Broadcast();
+	};
+
+	virtual void OnReturnToActorPool() override
+	{
+		INActorPoolItem::OnReturnToActorPool();
+		OnReturnToActorPoolEvent.Broadcast();
+	};
+
+	virtual void OnSpawnedFromActorPool() override
+	{
+		INActorPoolItem::OnSpawnedFromActorPool();
+		OnSpawnedFromActorPoolEvent.Broadcast();
+	};
+
+public:	
+	UPROPERTY(BlueprintAssignable)
+	FOnActorPoolDelegate OnCreatedByActorPoolEvent;
+	
+	UPROPERTY(BlueprintAssignable)
+	FOnActorPoolDelegate OnSpawnedFromActorPoolEvent;
+	
+	UPROPERTY(BlueprintAssignable)
+	FOnActorPoolDelegate OnReturnToActorPoolEvent;
+	
+	UPROPERTY(BlueprintAssignable)
+	FOnActorPoolDelegate OnDestroyedByActorPoolEvent;
 };